Creates platform specific config, main modules
This commit separates platform specific config and plugin serving code
from the generic one. It will aid the development of the Windows plugin.

* Extracts linux config from config.go into config_linux.go file.
* Extracts unix sock service from main.go into main_linux.go.
